Recoil fix (#106)
* Fixed recoil so it always goes back to where it started if no mouse movements were made.
Recoil is now calculated based on a second order polynomial. Coefficients determine the magnitude of the recoil and the time it takes to return to its original target after the recoil. These are constant for all guns at the moment but will be alterable through Blueprints next commit.

* Added horizontal recoil, and updated recoil values for all guns.

* Anchored HUD, crosshair and zoomed crosshair to the center of the screen properly.
Fixed sniper so that if you switch guns while zoomed in, the zoomed crosshair doesn't remain on the screen.

* Removed sniper zoom fix as its fixed already in dev branch